requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0
pypdf==4.0.1
youtube-transcript-api
Authlib
beautifulsoup4
